* 🚀 chore(README.md): update Go reference badge to v6 * 📝 docs(README.md): update code generator options documentation The Go reference badge has been updated to v6 to reflect the current version of the package. The documentation for the code generator options has been updated to reflect the new file structure and to provide more clarity on the available options. The `OptionWithSecret` and `OptionWithPeriod` options now have links to their respective files. The `OptionWithEmojies` option now links to the `emojies` package and provides instructions on how to add custom emojies. A new option, `OptionFromEncoded`, has been added to allow for the use of an exported generator.
* 📝 docs(README.md): update function signature and version history
The ValidateToken function signature has been updated to reflect the correct line numbers in the code. The ToEmojies function signature has also been updated to reflect the correct line numbers in the code. The version history has been updated to reflect the latest version, v6.0.0, and the previous version, v5.0.2, has been moved to the older section.

* 📝 docs(README.md): update completed tasks and add new ones
The completed tasks are marked with strikethrough. The new tasks include fixing a typo in the file path and updating the version of the package on pkg.go.dev.